Fix int overflow in hill climber sample counter sum
In determineAdjustment(), hitsInSample() + missesInSample() was computed
as int + int, which silently overflows when the true sum exceeds
Integer.MAX_VALUE. This occurs for caches with maximumSize >= ~107M where
sampleSize saturates at Integer.MAX_VALUE. The overflowed (negative) sum
permanently fails the requestCount < sampleSize check, preventing the
hill climber from adjusting the window/main partition and feeding a
negative hit rate into the step size calculation.

+++ caffeine/src/main/java/com/github/benmanes/caffeine/cache/BoundedLocalCache.java
@@ -1142,7 +1142,7 @@ abstract class BoundedLocalCache<K, V> extends BLCHeader.DrainStatusRef
       return;
     }
 
-    int requestCount = hitsInSample() + missesInSample();
+    long requestCount = (long) hitsInSample() + missesInSample();
     if (requestCount < frequencySketch().sampleSize) {
       return;
     }
